Dynamic X and Y setter

This function allows the dynamic adjustment of X and Y values in the makeMatrix function dependant on the desired group overlap. It uses a previously generated table from the overlapProbability function. The use of this function allows much faster generation of matrices with a specific group overlap than if the computer were to need to randomly try X and Y values until the overlap was generated. Y values are currently always (0,101) and X2 values are always x1 + 100. Is appropriate only to use when the points argument to makeMatrix is 102 and the by argument in 0.9. The returned X and Y values have been optimized to have the highest probability to generate the same overlap in both dimensions of the matrix i.e. when the makeMatrix argument dim is set to 2.

Arguments

desiredOverlap
Overlap for which X and Y values should be generated.

Examples

XY <- dynamicXY(desiredOverlap = 0)
mat <- makeMatrix(XY[[1]], XY[[2]], dim=2, points=102)
groups <- makeGroups(mat, names=c("grp1", "grp2"))
calculateOverlap(mat[[1]][ ,1], groups) #overlap dimension 1

#> [1] 0

calculateOverlap(mat[[1]][ ,2], groups) #overlap dimension 2

#> [1] 0