Zero or Hundred Matrix

Creates the zero or hundred dataset.

Arguments

overlap
The desired numerical overlap to be generated. Can be either 0 or 100.

dim
The number of desired spatial dimensions for each groups points to occupy.

reps
The number of times a dataset for a specific number of points should be generated.

p
The interval of points per group for which datasets should be generated.

verbose
Logical indicating if the program should report status updates.

save
Logical indicating if the output should be saved to the data folder of the current directory.

Details

Creates one of two datasets; one where the 2 groups have 0 percent overlap with each other or one where the two groups have 100 percent overlap with each other. 0 percent, in this context, is defined as no datapoints from one group occupy the same space as the points of the other group. 100 percent overlap is defined as one group completley containing the other group within the space that it occupies. For ClusterSignificance testing this function was run with the default paramaters for overlaps 0 and 100.

Examples

zeroOrHundred(overlap=0, reps=2, p=5:6, save=FALSE)

#> [1] "points: 5"
#> [1] "points: 6"
#> [1] "rep: 1"
#> [1] "rep: 2"

#> $`5.points`
#> , , 1.repitition
#> 
#>        [,1]  [,2]
#>  [1,] 159.5 176.6
#>  [2,] 101.0 128.9
#>  [3,] 129.8 110.9
#>  [4,] 128.0 193.7
#>  [5,] 179.3 140.6
#>  [6,]  26.1  45.9
#>  [7,]  72.9  97.2
#>  [8,]  90.0  57.6
#>  [9,]  93.6  94.5
#> [10,]   6.3  74.7
#> 
#> , , 2.repitition
#> 
#>        [,1]  [,2]
#>  [1,] 120.8 128.0
#>  [2,] 169.4 149.6
#>  [3,] 191.0 132.5
#>  [4,] 128.0 155.0
#>  [5,] 110.9 126.2
#>  [6,]  71.1  19.8
#>  [7,]  53.1  38.7
#>  [8,]  80.1  88.2
#>  [9,]  94.5  54.9
#> [10,]  10.8  81.9
#> 
#> 
#> $`6.points`
#> , , 1.repitition
#> 
#>          [,1]    [,2]
#>  [1,] 159.500 176.600
#>  [2,] 101.000 128.900
#>  [3,] 129.800 110.900
#>  [4,] 128.000 193.700
#>  [5,] 179.300 140.600
#>  [6,] 117.804 139.494
#>  [7,]  26.100  45.900
#>  [8,]  72.900  97.200
#>  [9,]  90.000  57.600
#> [10,]  93.600  94.500
#> [11,]   6.300  74.700
#> [12,]  81.559  33.101
#> 
#> , , 2.repitition
#> 
#>          [,1]    [,2]
#>  [1,] 120.800 128.000
#>  [2,] 169.400 149.600
#>  [3,] 191.000 132.500
#>  [4,] 128.000 155.000
#>  [5,] 110.900 126.200
#>  [6,] 117.804 139.494
#>  [7,]  71.100  19.800
#>  [8,]  53.100  38.700
#>  [9,]  80.100  88.200
#> [10,]  94.500  54.900
#> [11,]  10.800  81.900
#> [12,]  81.559  33.101
#> 
#>